Move common socket_handler code to module_utils.
